Separated next code and style code into different files
