+++ src/components/Header/NavContents/NavContents.styles.jsx
@@ -1,0 +1,10 @@
+import Link from "next/link";
+import styled from "styled-components";
+
+export const NavContent = styled.li`
+	padding: 5px 20px 10px 0;
+`;
+
+export const NavContentLink = styled(Link)`
+	font-size: 1.5rem;
+`;
