+++ src/components/Portfolio/ProjectTitle/ProjectTitle.jsx
@@ -1,0 +1,10 @@
+import { Technical, Title } from "./ProjectTitle.styles";
+
+export function ProjectTitle({ title, technical }) {
+	return (
+		<>
+			<Title>{title}</Title>
+			<Technical>{technical}</Technical>
+		</>
+	);
+}
