+++ src/components/Portfolio/ProjectTitle/ProjectTitle.styles.jsx
@@ -1,24 +1,15 @@
 import { SMARTPHONE } from "@/constants";
 import styled from "styled-components";
 
-const Title = styled.h2`
+export const Title = styled.h2`
 	font-weight: 600;
 	margin-bottom: 20px;
 `;
 
-const Technical = styled.p`
+export const Technical = styled.p`
 	color: var(--color-orange);
 
 	@media screen and (max-width: ${SMARTPHONE}) {
 		font-size: 0.8rem;
 	}
 `;
-
-export function ProjectTitle({ title, technical }) {
-	return (
-		<>
-			<Title>{title}</Title>
-			<Technical>{technical}</Technical>
-		</>
-	);
-}
