+++ src/pages/index.js
@@ -10,12 +10,11 @@ export default function Home() {
 				<meta name="description" content="Generated by create next app" />
 				<meta name="viewport" content="width=device-width, initial-scale=1" />
 			</Head>
-			<main>
-				<GlobalStyle />
-				<Header />
-				<PageLayout />
-				<SocialSideBar />
-			</main>
+
+			<GlobalStyle />
+			<Header />
+			<PageLayout />
+			<SocialSideBar />
 		</>
 	);
 }
